refactor: redo quite a bit of auto
