--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/autonomus/BlueSideAutoAuto.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.autoextend.autonomus;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-import org.firstinspires.ftc.teamcode.BlueSideAuto;
-
-@Autonomous(name="Auto: Blue Side", group = "Competition Ready")
-public class BlueSideAutoAuto extends BlueSideAuto {
-}
