--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/autonomus/PrimaryAutoModeAutonomus.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.autoextend.autonomus;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-import org.firstinspires.ftc.teamcode.PrimaryAutoModeClass;
-
-@Autonomous(name="Primary Auto Mode", group = "Competition Ready")
-public class PrimaryAutoModeAutonomus extends PrimaryAutoModeClass {
-}
