--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/autonomus/RedSideAutoAuto.java
@@ -1,10 +1,0 @@
-//package org.firstinspires.ftc.teamcode.autoextend.autonomus;
-//
-//import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-//
-//import org.firstinspires.ftc.teamcode.PrimaryAutoModeClass;
-//import org.firstinspires.ftc.teamcode.RedSideAuto;
-//
-//@Autonomous(name="RedSideAuto", group = "Competition Ready")
-//public class RedSideAutoAuto extends RedSideAuto {
-//}
