--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/teleops/BlueSideAutoTeleop.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.autoextend.teleops;
-
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-import org.firstinspires.ftc.teamcode.BlueSideAuto;
-import org.firstinspires.ftc.teamcode.PrimaryAutoModeClass;
-
-@TeleOp(name="BlueSideDebugging", group = "Competition Ready")
-public class BlueSideAutoTeleop extends BlueSideAuto {
-}
