--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/teleops/PrimaryAutoModeTeleop.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.autoextend.teleops;
-
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-import org.firstinspires.ftc.teamcode.PrimaryAutoModeClass;
-
-@TeleOp(name="Primary Auto Mode Debug", group = "Competition Ready")
-public class PrimaryAutoModeTeleop extends PrimaryAutoModeClass {
-}
