--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoextend/teleops/RedSideAutoTeleop.java
@@ -1,9 +1,0 @@
-//package org.firstinspires.ftc.teamcode.autoextend.teleops;
-//
-//import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-//
-//import org.firstinspires.ftc.teamcode.RedSideAuto;
-//
-//@TeleOp(name="RedSideDeugging", group = "Competition Ready")
-//public class RedSideAutoTeleop extends RedSideAuto {
-//}
