+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/opmodes/autonomus/BlueSideAutoAuto.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.autos.opmodes.autonomus;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.autos.classes.BlueAuto;
+
+@Autonomous(name="Auto: Blue Side", group = "Competition Ready")
+public class BlueSideAutoAuto extends BlueAuto {
+}
