+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/opmodes/autonomus/RedSideAutoAuto.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.autos.opmodes.autonomus;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.autos.classes.RedAuto;
+
+@Autonomous(name="Auto: Red Side", group = "Competition Ready")
+public class RedSideAutoAuto extends RedAuto {
+}
