+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/opmodes/teleops/BlueSideAutoTeleop.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.autos.opmodes.teleops;
+
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+import org.firstinspires.ftc.teamcode.autos.classes.BlueAuto;
+
+@TeleOp(name="Auto: Blue Side (DEBUG)", group = "Debugging")
+public class BlueSideAutoTeleop extends BlueAuto {
+}
