+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/opmodes/teleops/RedSideAutoTeleop.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.autos.opmodes.teleops;
+
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+import org.firstinspires.ftc.teamcode.autos.classes.RedAuto;
+
+@TeleOp(name="Auto: Red Side (DEBUG)", group = "Debugging")
+public class RedSideAutoTeleop extends RedAuto {
+}
